upd build
rm compile warnings

+++ File.cpp
@@ -19,7 +19,7 @@ struct File::Private {
   std::fstream f_;
   std::ios::openmode m_;
   std::string fn_;
-  std::uintmax_t fs_;
+  std::size_t fs_;
 };
 
 File::File(const std::string &fn) {
@@ -64,7 +64,7 @@ void File::remove() {
   std::filesystem::remove(private_->fn_);
 }
 
-std::uintmax_t File::size() { return private_->fs_; }
+std::size_t File::size() { return private_->fs_; }
 
 bool File::exists() { return std::filesystem::exists(private_->fn_); }
